Move `LinkButton` to navigation group
* It should only be used in "navigation-like" contexts
* Its styling and props are aligned with "Link", which is a navigation
  component
