+++ src/components/input/index.js
@@ -3,4 +3,3 @@ export { default as ButtonUnstyled } from './ButtonUnstyled';
 export { default as IconButton } from './IconButton';
 export { default as Input } from './Input';
 export { default as InputGroup } from './InputGroup';
-export { default as LinkButton } from './LinkButton';
