+++ src/components/navigation/index.js
@@ -1,2 +1,3 @@
 export { default as Link } from './Link';
 export { default as LinkUnstyled } from './LinkUnstyled';
+export { default as LinkButton } from './LinkButton';
