+++ src/next.js
@@ -10,7 +10,6 @@ export {
   Button,
   ButtonUnstyled,
   IconButton,
-  LinkButton,
   Input,
   InputGroup,
 } from './components/input';
@@ -22,4 +21,4 @@ export {
   CardActions,
   Panel,
 } from './components/layout';
-export { Link, LinkUnstyled } from './components/navigation/';
+export { Link, LinkUnstyled, LinkButton } from './components/navigation/';
