+++ src/pattern-library/routes.js
@@ -16,12 +16,12 @@ import ButtonsPage from './components/patterns/input/ButtonPage';
 import IconButtonPage from './components/patterns/input/IconButtonPage';
 import InputPage from './components/patterns/input/InputPage';
 import InputGroupPage from './components/patterns/input/InputGroupPage';
-import LinkButtonPage from './components/patterns/input/LinkButtonPage';
 
 import CardPage from './components/patterns/layout/CardPage';
 import PanelPage from './components/patterns/layout/PanelPage';
 
 import LinkPage from './components/patterns/navigation/LinkPage';
+import LinkButtonPage from './components/patterns/navigation/LinkButtonPage';
 
 // Legacy pattern-library pages
 import LayoutFoundations from './components/patterns/LayoutFoundations';
@@ -225,12 +225,6 @@ const routes = [
     route: '/input-iconbutton',
   },
   {
-    title: 'LinkButton',
-    group: 'input',
-    component: LinkButtonPage,
-    route: '/input-linkbutton',
-  },
-  {
     title: 'Panel',
     group: 'layout',
     component: PanelPage,
@@ -260,6 +254,12 @@ const routes = [
     group: 'navigation',
     route: '/navigation-link',
     component: LinkPage,
+  },
+  {
+    title: 'LinkButton',
+    group: 'navigation',
+    component: LinkButtonPage,
+    route: '/input-linkbutton',
   },
 ];
